feat: unit tests for driver-service

+++ driver-service/src/main/java/com/modsen/driver/repository/DriverRepository.java
@@ -14,7 +14,7 @@ public interface DriverRepository extends JpaRepository<Driver, Long> {
 
     Optional<Driver> findFirstByDriverStatusAndIdIsNotIn(DriverStatus status, List<Long> driversToExcludeIdList);
 
-    int countAllByDriverStatus(DriverStatus driverStatus);
+    long countAllByDriverStatus(DriverStatus driverStatus);
 
     boolean existsByLicenceNumber(String licenceNumber);
